+++ bagua/user_prefs.py
@@ -11,6 +11,7 @@ METHOD_NUM_TO_KEY = {
     "4": "number",
     "5": "manual",
     "6": "yarrow",
+    "7": "character",
 }
 METHOD_KEY_TO_NUM = {v: k for k, v in METHOD_NUM_TO_KEY.items()}